Changed to make Java 1.5 compliant
Previously spring-test-mvc required JDK 1.6 to build successfully. Since the pom was configured for JDK 1.5 and most Spring projects still support 1.5 I assume this was not intentional. The problem was that spring-text-mvc used String.getBytes(Charset) and the XML API's (in the tests) which are not available in JDK 1.5.

Now String.getBytes(String) is used and a test dependency on jaxb-api was added. These changes make the project truely JDK 1.5 compatible.