Update README.md to include scope test in dependency declaration
Previously the dependency declarations in the README.md did not include the test scope. While users should be able to update this on their own, it probably will help some copy paste errors since most projects will use this as a test dependency.

I have added the test scope to the dependency declarations.