MockMvcBuilders return concrete implementation
Currently MockMvcBuilders.webApplicationContextSetup returns
MockMvcBuilder rather than InitializedContextMockMvcBuilder. This is
problematic for anyone wanting to add a Filter which is defined on the
AbstractMockMvcBuilder.

This update changes MockMvcBuilders.webApplicationContextSetup to return
InitializedContextMockMvcBuilder which allows adding a Filter and is more
consistent with the other factory methods.MockMvcBuilders should return InitializedContextMockMvcBuilder